Refactor Inset calculation to private extension
